nicoles changes
